code components fixes
