+++ packages/toolpad-app/src/components/Studio.tsx
@@ -11,15 +11,17 @@ export interface EditorProps {
 
 export default function Editor({ basename }: EditorProps) {
   return (
-    <BrowserRouter basename={basename}>
-      <Routes>
-        <Route>
-          <Route path="/" element={<Home />} />
-          <Route path="/app/:appId/editor/*" element={<AppEditor />} />
-          <Route path="/app/:appId/releases" element={<Releases />} />
-          <Route path="/app/:appId/releases/:version" element={<Release />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
+    <React.Suspense fallback="loading...">
+      <BrowserRouter basename={basename}>
+        <Routes>
+          <Route>
+            <Route path="/" element={<Home />} />
+            <Route path="/app/:appId/editor/*" element={<AppEditor />} />
+            <Route path="/app/:appId/releases" element={<Releases />} />
+            <Route path="/app/:appId/releases/:version" element={<Release />} />
+          </Route>
+        </Routes>
+      </BrowserRouter>
+    </React.Suspense>
   );
 }
