+++ packages/toolpad-app/src/runtime/createCodeComponent.tsx
@@ -3,36 +3,7 @@ import * as React from 'react';
 import * as ReactIs from 'react-is';
 import { transform } from 'sucrase';
 import { findImports, isAbsoluteUrl } from '../utils/strings';
-
-async function resolveValues(input: Map<string, Promise<unknown>>): Promise<Map<string, unknown>> {
-  const resolved = await Promise.all(input.values());
-  return new Map(Array.from(input.keys(), (key, i) => [key, resolved[i]]));
-}
-
-async function createRequire(urlImports: string[]) {
-  const modules = await resolveValues(
-    new Map<string, any>([
-      ['react', import('react')],
-      ['react-dom', import('react-dom')],
-      ['@mui/toolpad-core', import(`@mui/toolpad-core`)],
-      ['@mui/material', import('@mui/material')],
-      ['@mui/material/Button', import('@mui/material/Button')],
-      // ... TODO: All @mui/material imports + custom solution for icons
-      ...urlImports.map((url) => [url, import(/* webpackIgnore: true */ url)] as const),
-    ]),
-  );
-
-  const require = (moduleId: string): unknown => {
-    const module = modules.get(moduleId);
-    if (module && typeof module === 'object') {
-      // ESM interop
-      return { ...module, __esModule: true };
-    }
-    throw new Error(`Can't resolve module "${moduleId}"`);
-  };
-
-  return require;
-}
+import createRequire from './createRequire';
 
 function ensureToolpadComponent<P>(Component: React.ComponentType<P>): ToolpadComponent<P> {
   if ((Component as any)[TOOLPAD_COMPONENT]) {
